+++ declafe/feature_gen/unary/times/MinuteFeature.py
@@ -13,7 +13,7 @@ class MinuteFeature(UnaryFeature):
   def gen_unary(self, ser: np.ndarray) -> np.ndarray:
     gen = np.frompyfunc(
         lambda x: datetime.utcfromtimestamp(x / 1000_000_000).minute, 1, 1)
-    return gen(ser)
+    return gen(ser).astype(np.int64)
 
   @property
   def name(self) -> str:
